progress-3

+++ src/Database/Component/DesignPatterm/IQueryType.java
@@ -10,18 +10,18 @@ package Database.Component.DesignPatterm;
  * @author Alim
  */
 public interface IQueryType {
-    int WHERE_AND       =   0;
-    int WHERE_OR        =   1;
-    int STARTS_WITH     =   2;
-    int CONTAINS        =   3;
-    int ENDS_WITH       =   4;
-    int GROUP_BY        =   5;
-    int ORDER_BY        =   6;
-    int MIN             =   7;
-    int MAX             =   8;
-    int AVG             =   9;
-    int COUNT           =   10;
-    int COMMAND         =   11;
+    byte WHERE_AND       =   0;
+    byte WHERE_OR        =   1;
+    byte STARTS_WITH     =   2;
+    byte CONTAINS        =   3;
+    byte ENDS_WITH       =   4;
+    byte GROUP_BY        =   5;
+    byte ORDER_BY        =   6;
+    byte MIN             =   7;
+    byte MAX             =   8;
+    byte AVG             =   9;
+    byte COUNT           =   10;
+    byte COMMAND         =   11;
     
 }
